@@ -4,7 +4,7 @@
   "license": "MIT",
   "scripts": {
     "ng": "nx",
-    "postinstall": "node ./decorate-angular-cli.js && ngcc --properties es2015 browser module main",
+    "postinstall": "node ./decorate-angular-cli.js && ngcc --properties es2020 browser module main",
     "start": "nx serve",
     "start:client": "yarn start client",
     "start:api": "yarn start api",
@@ -18,24 +18,24 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "~13.3.0",
-    "@angular/cdk": "^13.3.7",
-    "@angular/common": "~13.3.0",
-    "@angular/compiler": "~13.3.0",
-    "@angular/core": "~13.3.0",
-    "@angular/flex-layout": "^13.0.0-beta.38",
-    "@angular/forms": "~13.3.0",
-    "@angular/material": "^14.0.0",
-    "@angular/platform-browser": "~13.3.0",
-    "@angular/platform-browser-dynamic": "~13.3.0",
-    "@angular/router": "~13.3.0",
+    "@angular/animations": "14.0.0",
+    "@angular/cdk": "14.0.1",
+    "@angular/common": "14.0.0",
+    "@angular/compiler": "14.0.0",
+    "@angular/core": "14.0.0",
+    "@angular/flex-layout": "13.0.0-beta.38",
+    "@angular/forms": "14.0.0",
+    "@angular/material": "14.0.1",
+    "@angular/platform-browser": "14.0.0",
+    "@angular/platform-browser-dynamic": "14.0.0",
+    "@angular/router": "14.0.0",
     "@nestjs/apollo": "^10.0.12",
     "@nestjs/common": "^8.4.4",
     "@nestjs/core": "^8.0.0",
     "@nestjs/cqrs": "^8.0.4",
     "@nestjs/graphql": "^10.0.12",
     "@nestjs/platform-express": "^8.0.0",
-    "@nrwl/angular": "14.1.4",
+    "@nrwl/angular": "14.2.1",
     "@prisma/client": "^3.14.0",
     "angular-material-source": "angular/material",
     "apollo-server-express": "^3.7.0",
@@ -48,39 +48,39 @@
     "zone.js": "~0.11.4"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~13.3.0",
-    "@angular-eslint/eslint-plugin": "~13.1.0",
-    "@angular-eslint/eslint-plugin-template": "~13.1.0",
-    "@angular-eslint/template-parser": "~13.1.0",
-    "@angular/cli": "~13.3.0",
-    "@angular/compiler-cli": "~13.3.0",
-    "@angular/language-service": "~13.3.0",
+    "@angular-devkit/build-angular": "14.0.0",
+    "@angular-eslint/eslint-plugin": "13.2.1",
+    "@angular-eslint/eslint-plugin-template": "13.2.1",
+    "@angular-eslint/template-parser": "13.2.1",
+    "@angular/cli": "~14.0.0",
+    "@angular/compiler-cli": "14.0.0",
+    "@angular/language-service": "14.0.0",
     "@nestjs/schematics": "^8.0.0",
     "@nestjs/testing": "^8.0.0",
-    "@nrwl/cli": "14.1.4",
-    "@nrwl/cypress": "14.1.4",
-    "@nrwl/eslint-plugin-nx": "14.1.4",
-    "@nrwl/jest": "14.1.4",
-    "@nrwl/linter": "14.1.4",
-    "@nrwl/nest": "^14.1.5",
-    "@nrwl/node": "14.1.5",
-    "@nrwl/workspace": "14.1.4",
+    "@nrwl/cli": "14.2.1",
+    "@nrwl/cypress": "14.2.1",
+    "@nrwl/eslint-plugin-nx": "14.2.1",
+    "@nrwl/jest": "14.2.1",
+    "@nrwl/linter": "14.2.1",
+    "@nrwl/nest": "14.2.1",
+    "@nrwl/node": "14.2.1",
+    "@nrwl/workspace": "14.2.1",
     "@types/jest": "27.4.1",
     "@types/node": "16.11.7",
-    "@typescript-eslint/eslint-plugin": "~5.18.0",
-    "@typescript-eslint/parser": "~5.18.0",
+    "@typescript-eslint/eslint-plugin": "5.24.0",
+    "@typescript-eslint/parser": "5.24.0",
     "cypress": "^9.1.0",
     "docker": "^1.0.0",
-    "eslint": "~8.12.0",
+    "eslint": "8.15.0",
     "eslint-config-prettier": "8.1.0",
     "eslint-plugin-cypress": "^2.10.3",
     "jest": "27.5.1",
-    "jest-preset-angular": "11.1.1",
-    "nx": "14.1.4",
-    "prettier": "^2.5.1",
+    "jest-preset-angular": "11.1.2",
+    "nx": "14.2.1",
+    "prettier": "2.6.2",
     "prisma": "^3.14.0",
     "ts-jest": "27.1.4",
-    "ts-node": "^10.7.0",
-    "typescript": "~4.6.2"
+    "ts-node": "10.8.1",
+    "typescript": "4.7.3"
   }
 }
