@@ -7,8 +7,20 @@ import { Observable, observable } from 'rxjs';
 })
 
 export class AdminService {
-  constructor(private http:HttpClient){}
 
+  company:any
+  constructor(private http:HttpClient){
+
+  }
+
+  getCompany(companyName:string)
+  {
+    const Query='query{GetCompanyQuery(name:"iCreateSoftware"){id,company_name,employees{name,surname},admins{name,surname},projects{project_name}}}'
+    this.http.post<any>("https://localhost:3333/graphql",JSON.stringify({ query: Query})).subscribe(data=>
+    {
+      this.company=data;
+    })
+  }
   getAdmins(companyName:string)
   {
     //
