@@ -8,8 +8,8 @@ import {MatSidenavModule} from "@angular/material/sidenav";
 import {MatToolbarModule} from "@angular/material/toolbar";
 import { Router } from '@angular/router';
 import {IndividualRoutingModule} from "./ind-routing.module";
-import { IndividualProfileComponent } from './individual-profile/individual-profile.component';
-import { DashboardComponent } from './dashboard/dashboard.component';
+//import { IndividualProfileComponent } from './individual-profile/individual-profile.component';
+//import { DashboardComponent } from './dashboard/dashboard.component';
 import {ClientAdminFeatureModule} from "@team-utilisation-monitor/client/admin/feature";
 import {MatGridListModule} from "@angular/material/grid-list";
 import {MatCardModule} from "@angular/material/card";
@@ -38,9 +38,7 @@ import {MatSelectModule} from "@angular/material/select";
     MatSelectModule,
   ],
   declarations: [
-    IndividualHomePageComponent,
-    IndividualProfileComponent,
-    DashboardComponent
+    IndividualHomePageComponent
   ],
   exports: [IndividualRoutingModule],
 })
